requirements.txt
flask
flask_sqlalchemy
numpy
pillow
tensorflow-cpu==2.13.0
gunicorn
werkzeug